cms_backend/requirements.txt
Django==5.2.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
requests==2.32.4

